refactor(client): merge status bar into header, fix sub-panel toggle DRY
Eliminates StatusBar.tsx — connection dot, theme name, and panel
toggle icons now live in the header. Saves 24px of vertical space.
Burger remains mobile-only; panel toggles are desktop-only.

Extracts the triplicated "create-if-none, toggle-if-exists" sub-panel
logic into a single handleToggleSubPanel function in App.tsx.
commands.ts and useShortcuts.ts now call it via the toggleSubPanel dep
instead of reimplementing the conditional.
